Interpreter: Rename namespace after being a submodule
