+++ include/core/interpreter/vm_register.hpp
@@ -9,7 +9,7 @@
 
 namespace dvm {
     namespace core {
-        namespace vm {
+        namespace interpreter {
             namespace registers {
                 class vm_context;
