+++ include/core/object/vm_class.hpp
@@ -1,0 +1,7 @@
+//
+// Created by kiva on 2017/8/31.
+//
+#pragma once
+
+
+
